# elementwise_add_grad Op optimization 