<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
| 输入(x_shape)/y_shape | data type | develop|this PR|
|---|---|---|---|
| [32, 128, 728]/ [728] | float16 |88.75us |16.78us|
|  [32, 128, 3072] /[3072]| float16 |136.39us|41.96us|