@wangchaochaohu 2020-12-14
> OP benchmark CI的个别配置是耗时增加的。

![image](https://user-images.githubusercontent.com/13411999/102087138-1119a380-3e54-11eb-892f-45bf9e32a4e8.png)
![image](https://user-images.githubusercontent.com/13411999/102087157-170f8480-3e54-11eb-8c18-607ee79f15ba.png)
这两个配置在后续PR优化